+++ agent/backend/src/main/java/eu/bbmri_eric/quality/agent/dataquality/impl/UpdateCheck.java
@@ -85,7 +85,7 @@ class UpdateCheck implements DataQualityCheck {
   }
 
   @Override
-  public Float getEpsilonBudget() {
+  public Double getEpsilonBudget() {
     return config.getEpsilonBudget();
   }
